+++ packages/util/test/promise.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, it } from 'vitest';
-import { sleepInMilliseconds, sleepInSeconds } from '../src/promise';
+import {
+	sequentialPromise,
+	sleepInMilliseconds,
+	sleepInSeconds,
+} from '../src/promise';
 
 describe('A promise helper', () => {
 	it('should sleep for a given second/millisecond', async () => {
@@ -20,5 +24,21 @@ describe('A promise helper', () => {
 		});
 
 		expect(anotherValue).toBe('hi');
+	});
+
+	it('should execute an array of asynchronous functions in sequential order', async () => {
+		const indexes = Array.from({ length: 5 }, (_, index) => {
+			return index;
+		});
+
+		const result = await sequentialPromise(
+			indexes.map((index) => {
+				return async () => {
+					return index;
+				};
+			})
+		);
+
+		expect(result).toEqual(indexes);
 	});
 });